#81 Refactor commands.py to core part
